fileSize 류 필드의 데이터 타입을 long형으로 변경
git-svn-id: https://skvalley.com/svn/repos/storeplatform/trunk/sac/storeplatform-sac@61553 01f5c84a-aead-4b29-b0f1-7c54dbeec05f

+++ src/main/java/com/skplanet/storeplatform/sac/display/app/vo/AppDetail.java
@@ -23,7 +23,7 @@ public class AppDetail extends CommonInfo {
     private String apkPkgNm;
     private String apkVer;
     private String apkVerNm;
-    private Integer fileSize;
+    private Long fileSize;
     private String vmVer;
     private String sellerMbrNo;
     private String prodGbn;
@@ -113,11 +113,11 @@ public class AppDetail extends CommonInfo {
         this.prodStatusCd = prodStatusCd;
     }
 
-    public Integer getFileSize() {
+    public Long getFileSize() {
         return fileSize;
     }
 
-    public void setFileSize(Integer fileSize) {
+    public void setFileSize(Long fileSize) {
         this.fileSize = fileSize;
     }
