+++ src/main/java/com/skplanet/storeplatform/sac/display/response/AppInfoGeneratorImpl.java
@@ -74,7 +74,7 @@ public class AppInfoGeneratorImpl implements AppInfoGenerator {
 	}
 
 	@Override
-	public App generateApp(String aid, String packageNm, String apkVer, String prodVer, Integer fileSize,
+	public App generateApp(String aid, String packageNm, String apkVer, String prodVer, Long fileSize,
 			String supportOs, String scid, String filePath) {
 		App app = new App();
 		app.setAid(aid);
